WD-290: Provide support for V4
