+++ package.json
@@ -1,42 +1,34 @@
 {
-  "name": "@mondago/strapi-provider-email-msgraph",
-  "version": "1.0.1",
-  "description": "Microsoft Graph email provider plugin for Strapi 3",
-  "main": "./lib",
+  "name": "strapi-provider-email-msgraph",
+  "version": "1.0.0",
+  "description": "Emails via Microsoft Graph",
+  "main": "./dist",
+  "types": "./dist/index.d.ts",
   "directories": {
-    "lib": "lib"
-  },
-  "dependencies": {
-    "@azure/identity": "^3.0.0",
-    "@microsoft/microsoft-graph-client": "^3.0.2",
-    "isomorphic-fetch": "^3.0.0"
+    "dist": "./dist"
   },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "format:check": "prettier --check src/**/*.{ts,tsx}",
+    "format": "prettier --write src/**/*.{ts,tsx}",
+    "build": "tsc",
+    "lint": "eslint .",
+    "clean": "rimraf ./dist"
   },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/mondago/strapi-provider-email-msgraph.git"
+  "engines": {
+    "node": ">=16.0.0 <=20.x.x",
+    "npm": ">=6.0.0"
   },
-  "keywords": [
-    "Strapi",
-    "Email",
-    "Provider",
-    "Strapi 3",
-    "Graph",
-    "Microsoft"
-  ],
-  "author": "Mondago Ltd",
-  "contributors": [
-    "Daniel Yates"
-  ],
-  "license": "MIT",
-  "homepage": "https://github.com/mondago/strapi-provider-email-msgraph#readme",
-  "strapi": {
-    "isProvider": true
+  "devDependencies": {
+    "prettier": "^3.0.3",
+    "@types/isomorphic-fetch": "^0.0.36",
+    "rimraf": "^5.0.1",
+    "typescript": "^5.2.2",
+    "eslint": "^8.48.0",
+    "@tsconfig/node16": "^16.1.1"
   },
-  "bugs": {
-    "url": "https://github.com/mondago/strapi-provider-email-msgraph/issues"
-  },
-  "devDependencies": {}
+  "dependencies": {
+    "@azure/identity": "^3.3.0",
+    "@microsoft/microsoft-graph-client": "^3.0.6",
+    "isomorphic-fetch": "^3.0.0"
+  }
 }
